Removed Id from fieldMapping
Since the Id is used internally as the nodes path it is not necessary to
store it with the node - it is actually impossible to do it with a
predefined node type that does not allow arbitrary properties as
nt:file. There for we need a different mapping for Id (as it was before
with Path).