[WIP] Implement reorder
Reordering nodes is a necessary requirement. Doing this via reordering the children mapped in a document can not be done in a generic way since the children mapping allows to filter certain children and you might want to reorder children even when there is no child mapping.

The admin tree allows to move nodes and also for reordering. But PhpcrOdmTree displays all children in one list - you don't even know which mapping a child belongs to when there is more than one mapping.

So this method implements the reordering directly at the phpcr level ignoring child and children mappings.

This currently works for a reordering in the admin tree but is not complete yet.