Getting phpcr bin usable when installed via Composer
The phpcr bin is now installed when phpcr-odm is installed
via Composer.

The entire process is a little less DRY as the root folder
bootstrap.php has been removed in favor of having that
code live both in bin/phpcr and in tests/bootstrap.php.
This was done to remove the requirement that the
cli-config.php scripts require bootstrap.php individually.