Set to references collection to empty array instead of null when empty
When setting to null this would cause the property to be deleted
which would result in the value being set to `null` in the object
when found later. I believe this is not to be expected as it would
otherwise require document classes to inspect their reference
collections before being used to see if it is "still" an array
or not.

This change required some changes to two tests that were expecting
the property to be deleted.

The rest of the tests from the entire suite ran fine w/ jackrabbit.